Screen with weather done

+++ app/src/main/java/com/romanenko/lew/progforcetask/model/IRepository.java
@@ -1,5 +1,6 @@
 package com.romanenko.lew.progforcetask.model;
 
+import com.romanenko.lew.progforcetask.entity.WeatherObject;
 import com.romanenko.lew.progforcetask.model.POJO.Weather;
 
 import io.reactivex.Observable;
@@ -7,5 +8,5 @@ import io.reactivex.Single;
 
 public interface IRepository {
 
-    Single<Weather> loadWeatherForecast5Days(float lat, float lon);
+    Single<Weather> loadWeatherForecast5Days(double lat, double lon);
 }
